Consolidate token and tokenlist modules
